Refactor stack_frame out of src/debug/mod.rs
