+++ debugger/src/debugger/session_data.rs
@@ -195,7 +195,7 @@ impl SessionData {
                 ),
                 debug_info,
                 core_peripherals: None,
-                stack_frames: Vec::<probe_rs::debug::StackFrame>::new(),
+                stack_frames: Vec::<probe_rs::debug::stack_frame::StackFrame>::new(),
                 breakpoints: Vec::<ActiveBreakpoint>::new(),
                 rtt_connection: None,
             })
